Move several functions to util module
